= Ganti dari tipe data dari int jadi double untuk variable amount + Call method addHistoy di credit dan deposit

+++ Documents/NetBeansProjects/APPLDBUP/src/History.java
@@ -10,14 +10,14 @@
  */
 public class History {
     private String keterangan;
-    private int amount;
+    private double amount;
 
-    public History(String keterangan, int amount) {
+    public History(String keterangan, double amount) {
         this.keterangan = keterangan;
         this.amount = amount;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
